+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fix spelling error in `window_size` field for`rate_limit` middleware.
 - Fix regex path order issue for routes.
 - Fix accuracy of `http_bifrost_request_duration` metric.
+- Fix dynamic service has been executed many times.
 
 ### **Refactor**